+++ tests/test_kernel32/test_events.py
@@ -7,11 +7,11 @@ from pywincffi.dev.testutil import TestCase
 from pywincffi.exceptions import WindowsAPIError, InputError
 from pywincffi.kernel32 import events  # used by mocks
 from pywincffi.kernel32 import (
-    CloseHandle, CreateEvent, OpenEvent)
+    CloseHandle, CreateEvent, OpenEvent, ResetEvent, WaitForSingleObject)
 
 
-# These tests cause TestPidExists to fail under Python 3.4 so for now
-# we skip these tests.  Because we're only testing CreateEvent, and
+# These tests cause TestPidExists and others to fail under Python 3.4 so for
+# now we skip these tests.  Because we're only testing CreateEvent, and
 # TestPidExists worked before TestCreateEvent exists, we'll skip these
 # for now.
 #  Traceback (most recent call last):
@@ -31,13 +31,30 @@ class TestCreateEvent(TestCase):
     def setUp(self):
         super(TestCreateEvent, self).setUp()
         if sys.version_info[0:2] == (3, 4):
-            self.skipTest("Not compatible with Python 3.4")
+            self.skipTest("Skipped on Python 3.4, see comments.")
 
     def test_create_event_valid_handle(self):
         handle = CreateEvent(False, False)
         CloseHandle(handle)  # will raise exception if the handle is invalid
 
+    def test_non_signaled(self):
+        handle = CreateEvent(False, False)
+        self.addCleanup(CloseHandle, handle)
+        _, library = dist.load()
+        self.assertEqual(
+            WaitForSingleObject(handle, 0), library.WAIT_TIMEOUT)
+
+    def test_signaled(self):
+        handle = CreateEvent(False, True)
+        self.addCleanup(CloseHandle, handle)
+        _, library = dist.load()
+        self.assertEqual(
+            WaitForSingleObject(handle, 0), library.WAIT_OBJECT_0)
+
     def test_creating_duplicate_event_does_not_raise_error(self):
+        if sys.version_info[0:2] == (3, 4):
+            self.skipTest("Not compatible with Python 3.4")
+
         # Windows raises set the last error to ERROR_ALREADY_EXISTS
         # if an event object with the same name already exists.  The
         # pywincffi API ignores this error and returns the handle
@@ -57,6 +74,9 @@ class TestCreateEvent(TestCase):
                 CreateEvent(False, False)
 
     def test_can_retrieve_named_event(self):
+        if sys.version_info[0:2] == (3, 4):
+            self.skipTest("Not compatible with Python 3.4")
+
         _, library = dist.load()
         name = "pywincffi-%s" % self.random_string(5)
         handle = CreateEvent(False, False, lpName=name)
@@ -67,3 +87,21 @@ class TestCreateEvent(TestCase):
     def test_check_lpeventattributes_type(self):
         with self.assertRaises(InputError):
             CreateEvent(False, False, lpEventAttributes="")
+
+
+class TestResetEvent(TestCase):
+    """
+    Tests for :func:`pywincffi.kernel32.ResetEvent`
+    """
+    def test_basic_reset(self):
+        handle = CreateEvent(True, True)
+        self.addCleanup(CloseHandle, handle)
+        ResetEvent(handle)
+
+    def test_resets_event(self):
+        handle = CreateEvent(True, True)
+        self.addCleanup(CloseHandle, handle)
+        ResetEvent(handle)
+
+        _, library = dist.load()
+        self.assertEqual(WaitForSingleObject(handle, 0), library.WAIT_TIMEOUT)