+++ docs/source/changelog.rst
@@ -18,6 +18,8 @@ Notable enhancements and changes are:
     * Several enhancements bringing :issue:`69` closer to closure.
     * Addition of the :func:`pywincffi.kernel32.events.CreateEvent` and
       :func:`pywincffi.kernel32.events.OpenEvent` functions in :issue:`70`.
+    * Addition of the :func:`pywincffi.kernel32.events.ResetEvent` in
+      :issue:`75`.
 
 0.2.0
 ~~~~~